# data/fossology-LicenseDb/cmd/laas/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service to host license information for other services to query over REST API.",
        "title": "laas (License as a Service) API",
        "contact": {
            "name": "FOSSology",
            "url": "https://fossology.org",
            "email": "fossology@fossology.org"
        },
        "license": {
            "name": "GPL-2.0-only",
            "url": "https://github.com/fossology/LicenseDb/blob/main/LICENSE"
        },
        "version": "0.0.9"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/audits": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get all audit records from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get audit records",
                "operationId": "GetAllAudit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit records",
                        "schema": {
                            "$ref": "#/definitions/models.AuditResponse"
                        }
                    },
                    "404": {
                        "description": "Not changelogs in DB",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/audits/{audit_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get a specific audit records by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get an audit record",
                "operationId": "GetAudit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit ID",
                        "name": "audit_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AuditResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid audit ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "404": {
                        "description": "No audit entry with given ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/audits/{audit_id}/changes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get changelogs of an audit record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get changelogs",
                "operationId": "GetChangeLogs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit ID",
                        "name": "audit_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ChangeLogResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid audit ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "404": {
                        "description": "No audit entry with given ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "500": {
                        "description": "unable to find changes",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/audits/{audit_id}/changes/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get a specific changelog of an audit record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get a changelog",
                "operationId": "GetChangeLogbyId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit ID",
                        "name": "audit_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Changelog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ChangeLogResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "404": {
                        "description": "No changelog with given ID found",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Fetches data to be displayed on the dashboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Fetches data to be displayed on the dashboard",
                "operationId": "GetDashboardData",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DashboardResponse"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Audit": {
            "type": "object",
            "properties": {
                "entity": {
                    "type": "object"
                },
                "id": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-01T18:10:25.00+05:30"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "OBLIGATION",
                        "LICENSE",
                        "USER"
                    ],
                    "example": "LICENSE"
                },
                "type_id": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                }
            }
        },
        "models.AuditResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Audit"
                    }
                },
                "paginationmeta": {
                    "$ref": "#/definitions/models.PaginationMeta"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "models.CategoryObligationCount": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "GENERAL"
                },
                "count": {
                    "type": "integer",
                    "example": 6
                }
            }
        },
        "models.ChangeLog": {
            "type": "object",
            "properties": {
                "audit_id": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                },
                "field": {
                    "type": "string",
                    "example": "text"
                },
                "id": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                },
                "old_value": {
                    "type": "string",
                    "example": "Old license text"
                },
                "updated_value": {
                    "type": "string",
                    "example": "New license text"
                }
            }
        },
        "models.ChangeLogResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChangeLog"
                    }
                },
                "paginationmeta": {
                    "$ref": "#/definitions/models.PaginationMeta"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.Dashboard": {
            "type": "object",
            "properties": {
                "category_obligation_frequency": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CategoryObligationCount"
                    }
                },
                "licenses_count": {
                    "type": "integer",
                    "example": 2
                },
                "monthly_license_changes_count": {
                    "type": "integer",
                    "example": 6
                },
                "obligations_count": {
                    "type": "integer",
                    "example": 7
                },
                "risk_license_frequency": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RiskLicenseCount"
                    }
                },
                "users_count": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.DashboardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Dashboard"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.LicenseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "path": {
                    "type": "string",
                    "example": "/api/v1/licenses"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-01T10:00:51+05:30"
                }
            }
        },
...
        "models.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "next": {
                    "type": "string",
                    "example": "/api/v1/licenses?limit=10\u0026page=11"
                },
                "page": {
                    "type": "integer",
                    "example": 10
                },
                "previous": {
                    "type": "string",
                    "example": "/api/v1/licenses?limit=10\u0026page=9"
                },
                "resource_count": {
                    "type": "integer",
                    "example": 200
                },
                "total_pages": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
...
        "models.RiskLicenseCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 6
                },
                "risk": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "fossy"
                },
                "id": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                },
                "subscribed": {
                    "type": "boolean"
                },
                "user_email": {
                    "type": "string",
                    "example": "fossy@org.com"
                },
                "user_level": {
                    "type": "string",
                    "example": "USER"
                },
                "user_name": {
                    "type": "string",
                    "example": "fossy"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token from /login endpoint. Enter the token with the `Bearer ` prefix, e.g. \\\"Bearer eyJhbGciOiJ.....\\\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}